util funcao para exibir imagens e funcao para abrir modal com imagem
