+++ app/Http/Controllers/UtilController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Classes\Util;
 use App\Models\FilesModel;
 use App\Models\FoldersModel;
 use Illuminate\Http\Request;
@@ -15,5 +16,10 @@ class UtilController extends Controller
         $files = FilesModel::where('parent_folder_id','=',$folder->id)->get();
 
         return view('local', ['folder' => $folder, 'files'=>$files, 'folders'=>$folders]);
+    }
+
+    public function returnTempLink($path)
+    {
+        return Util::displayImage($path);
     }
 }
